+++ src/instruments/options/black_scholes_merton.rs
@@ -15,7 +15,7 @@ use crate::instruments::options::TypeFlag;
 use crate::statistics::distributions::{Distribution, Gaussian};
 use crate::time::{DayCountConvention, DayCounter};
 
-use time::{Duration, OffsetDateTime};
+use time::OffsetDateTime;
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // STRUCTS, ENUMS, AND TRAITS
@@ -78,14 +78,10 @@ impl BlackScholesMerton {
 
     /// Generalised Black-Scholes European Option Price.
     pub fn price(&self) -> f64 {
-        let (S, K, v, r, b) = self.unpack();
+        let (S, K, _, r, b) = self.unpack();
 
         // Compute time to maturity.
-        let T = DayCounter::day_count_factor(
-            self.evaluation_date.unwrap_or(OffsetDateTime::now_utc()),
-            self.expiration_date,
-            &DayCountConvention::Actual365,
-        );
+        let T = self.year_fraction();
 
         let (d1, d2) = self.d1_d2();
 
@@ -99,16 +95,21 @@ impl BlackScholesMerton {
         }
     }
 
-    /// Compute d1 and d2.
-    pub fn d1_d2(&self) -> (f64, f64) {
-        let (S, K, v, r, b) = self.unpack();
-
-        // Compute time to maturity.
-        let T = DayCounter::day_count_factor(
+    // Compute the year fraction between two dates.
+    fn year_fraction(&self) -> f64 {
+        DayCounter::day_count_factor(
             self.evaluation_date.unwrap_or(OffsetDateTime::now_utc()),
             self.expiration_date,
             &DayCountConvention::Actual365,
-        );
+        )
+    }
+
+    // Compute d1 and d2.
+    fn d1_d2(&self) -> (f64, f64) {
+        let (S, K, v, _, b) = self.unpack();
+
+        // Compute time to maturity.
+        let T = self.year_fraction();
 
         let d1 = (1.0 / (v * T.sqrt())) * ((S / K).ln() + (b + 0.5 * v.powi(2)) * T);
         let d2 = d1 - v * T.sqrt();
@@ -129,14 +130,10 @@ impl BlackScholesMerton {
 
     /// Delta of generalised Black-Scholes European Option.
     pub fn delta(&self) -> f64 {
-        let (S, K, v, r, b) = self.unpack();
+        let (_, _, _, r, b) = self.unpack();
 
         // Compute time to maturity.
-        let T = DayCounter::day_count_factor(
-            self.evaluation_date.unwrap_or(OffsetDateTime::now_utc()),
-            self.expiration_date,
-            &DayCountConvention::Actual365,
-        );
+        let T = self.year_fraction();
 
         let d1 = self.d1_d2().0;
 
@@ -151,14 +148,10 @@ impl BlackScholesMerton {
     /// Vanna of generalised Black-Scholes European Option.
     /// Also known as DdeltaDvol.
     pub fn vanna(&self) -> f64 {
-        let (S, K, v, r, b) = self.unpack();
+        let (_, _, v, r, b) = self.unpack();
 
         // Compute time to maturity.
-        let T = DayCounter::day_count_factor(
-            self.evaluation_date.unwrap_or(OffsetDateTime::now_utc()),
-            self.expiration_date,
-            &DayCountConvention::Actual365,
-        );
+        let T = self.year_fraction();
 
         let (d1, d2) = self.d1_d2();
 
@@ -170,14 +163,10 @@ impl BlackScholesMerton {
     /// Charm of generalised Black-Scholes European Option.
     /// Also known as DdeltaDtime, delta decay or delta bleed.
     pub fn charm(&self) -> f64 {
-        let (S, K, v, r, b) = self.unpack();
+        let (_, _, v, r, b) = self.unpack();
 
         // Compute time to maturity.
-        let T = DayCounter::day_count_factor(
-            self.evaluation_date.unwrap_or(OffsetDateTime::now_utc()),
-            self.expiration_date,
-            &DayCountConvention::Actual365,
-        );
+        let T = self.year_fraction();
 
         let (d1, d2) = self.d1_d2();
 
@@ -198,35 +187,16 @@ impl BlackScholesMerton {
     /// Lambda of generalised Black-Scholes European Option.
     /// Also known as elasticity or leverage.
     pub fn lambda(&self) -> f64 {
-        let n = Gaussian::default();
-
-        let (S, K, v, r, b) = self.unpack();
-
-        let T = DayCounter::day_count_factor(
-            self.evaluation_date.unwrap_or(OffsetDateTime::now_utc()),
-            self.expiration_date,
-            &DayCountConvention::Actual365,
-        );
-
-        let (d1, d2) = self.d1_d2();
-
-        self.delta() * S / self.price()
+        self.delta() * self.underlying_price / self.price()
     }
 
     /// Gamma of generalised Black-Scholes European Option.
     /// Also known as convexity.
     pub fn gamma(&self) -> f64 {
         let n = Gaussian::default();
-
-        let (S, K, v, r, b) = self.unpack();
-
-        let T = DayCounter::day_count_factor(
-            self.evaluation_date.unwrap_or(OffsetDateTime::now_utc()),
-            self.expiration_date,
-            &DayCountConvention::Actual365,
-        );
-
-        let (d1, d2) = self.d1_d2();
+        let (S, _, v, r, b) = self.unpack();
+        let T = self.year_fraction();
+        let (d1, _) = self.d1_d2();
 
         ((b - r) * T).exp() * n.pdf(d1) / (S * v * T.sqrt())
     }
@@ -239,21 +209,8 @@ impl BlackScholesMerton {
     /// Zomma of generalised Black-Scholes European Option.
     /// Also known as DgammaDvol.
     pub fn zomma(&self) -> f64 {
-        let n = Gaussian::default();
-
-        let (S, K, v, r, b) = self.unpack();
-
-        let T = DayCounter::day_count_factor(
-            self.evaluation_date.unwrap_or(OffsetDateTime::now_utc()),
-            self.expiration_date,
-            &DayCountConvention::Actual365,
-        );
-
         let (d1, d2) = self.d1_d2();
-
-        let gamma = self.gamma();
-
-        gamma * ((d1 * d2 - 1.0) / v)
+        self.gamma() * ((d1 * d2 - 1.0) / self.volatility)
     }
 
     /// Zomma percent of generalised Black-Scholes European Option.
@@ -264,17 +221,9 @@ impl BlackScholesMerton {
     /// Speed of generalised Black-Scholes European Option.
     /// Also known as DgammaDspot.
     pub fn speed(&self) -> f64 {
-        let n = Gaussian::default();
-
-        let (S, K, v, r, b) = self.unpack();
-
-        let T = DayCounter::day_count_factor(
-            self.evaluation_date.unwrap_or(OffsetDateTime::now_utc()),
-            self.expiration_date,
-            &DayCountConvention::Actual365,
-        );
-
-        let (d1, d2) = self.d1_d2();
+        let (S, _, v, _, _) = self.unpack();
+        let T = self.year_fraction();
+        let (d1, _) = self.d1_d2();
 
         let gamma = self.gamma();
 
@@ -284,21 +233,51 @@ impl BlackScholesMerton {
     /// Colour of generalised Black-Scholes European Option.
     /// Also known as DgammaDtime.
     pub fn colour(&self) -> f64 {
-        let n = Gaussian::default();
-
-        let (S, K, v, r, b) = self.unpack();
-
-        let T = DayCounter::day_count_factor(
-            self.evaluation_date.unwrap_or(OffsetDateTime::now_utc()),
-            self.expiration_date,
-            &DayCountConvention::Actual365,
-        );
-
+        let (_, _, v, r, b) = self.unpack();
+        let T = self.year_fraction();
         let (d1, d2) = self.d1_d2();
 
         let gamma = self.gamma();
 
         gamma * (r - b + b * d1 / (v * T.sqrt()) + (1.0 - d1 * d2) / (2.0 * T))
+    }
+
+    /// Vega of generalised Black-Scholes European Option.
+    /// Also known as zeta.
+    pub fn vega(&self) -> f64 {
+        let (S, _, _, r, b) = self.unpack();
+        let T = self.year_fraction();
+        let (d1, _) = self.d1_d2();
+
+        let n = Gaussian::default();
+
+        S * ((b - r) * T).exp() * n.pdf(d1) * T.sqrt()
+    }
+
+    /// Vomma of generalised Black-Scholes European Option.
+    /// Also known as DvegaDvol.
+    pub fn vomma(&self) -> f64 {
+        let (d1, d2) = self.d1_d2();
+
+        self.vega() * d1 * d2 / self.volatility
+    }
+
+    /// Ultima of generalised Black-Scholes European Option.
+    /// Also known as DvommaDvol.
+    pub fn ultima(&self) -> f64 {
+        let (d1, d2) = self.d1_d2();
+
+        (self.vomma() / self.volatility) * (d1 * d2 - d1 / d2 + d2 / d1 - 1.0)
+    }
+
+    /// Vega Bleed of the generalised Black-Scholes European option.
+    /// Also known as DvegaDtime.
+    pub fn vega_bleed(&self) -> f64 {
+        let (_, _, v, r, b) = self.unpack();
+        let T = self.year_fraction();
+        let (d1, d2) = self.d1_d2();
+
+        self.vega() * (r - b + b * d1 / (v * T.sqrt()) - (d1 * d2 + 1.0) / (2.0 * T))
     }
 }
 
@@ -309,6 +288,7 @@ impl BlackScholesMerton {
 #[cfg(test)]
 mod tests_black_scholes_merton {
     use super::*;
+    use time::Duration;
 
     #[test]
     fn black_scholes_1973() {